requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
bleach==6.0.0
Django==3.2.25
django-filebrowser==4.0.3
django-grappelli==3.0.9
django-tinymce==3.6.1
Pillow==9.5.0
pytz==2024.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tinycss2==1.2.1
typing-extensions==4.7.1
webencodings==0.5.1
gunicorn>=20.0.4
django-environ==0.11.2
whitenoise==6.5.0
pre-commit
